typo

+++ usr/share/arcologout/Functions.py
@@ -171,7 +171,7 @@ def _get_logout():
         return "gnome-session-quit --logout --no-prompt"
     elif desktop in ("gnome-classic", "/usr/share/xsessions/gnome-classic"):
         return "gnome-session-quit --logout --no-prompt"
-    elfi desktop:
+    elif desktop:
 		return "pkill -KILL -u $USER"
     return None
